requirements.txt
Flask
tensorflow
pillow
numpy

